Feature/1/Seeds

+++ db/migrate/20190108220224_create_markers.rb
@@ -1,8 +1,8 @@
 class CreateMarkers < ActiveRecord::Migration[5.2]
   def change
     create_table :markers do |t|
-      t.integer :coord_lat
-      t.integer :coord_lng
+      t.decimal :coord_lat
+      t.decimal :coord_lng
       t.string :content
       t.string :title
       t.string :type
